+++ ProjectInventarisUAS/app/src/main/java/com/pplm/projectinventarisuas/ui/adminsection/item/ItemFragment.kt
@@ -5,7 +5,10 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import androidx.core.widget.doOnTextChanged
+import androidx.recyclerview.widget.GridLayoutManager
+import com.pplm.projectinventarisuas.data.model.ItemSummary
+import com.pplm.projectinventarisuas.utils.adapter.ItemSummaryAdapter
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -25,6 +28,7 @@ class ItemFragment : Fragment() {
     private val binding get() = _binding!!
     private lateinit var viewModel: ItemViewModel
     private lateinit var adapter: ItemAdapter
+    private lateinit var summaryAdapter: ItemSummaryAdapter
     private var isDialogVisible = false
 
     override fun onCreateView(
@@ -38,8 +42,17 @@ class ItemFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        setupSearchBar()
+        setupSwipeToRefresh()
         setupRecyclerView()
         setupViewModel()
+    }
+
+    private fun setupSearchBar() {
+        binding.etSearch.doOnTextChanged { text, _, _, _ ->
+            val query = text.toString()
+            viewModel.searchItems(query)
+        }
     }
 
     private fun setupRecyclerView() {
@@ -53,11 +66,23 @@ class ItemFragment : Fragment() {
     private fun setupViewModel() {
         val borrowingRepository = BorrowingRepository()
         val userRepository = UserRepository()
-        val viewModelFactory = ViewModelFactory(ItemRepository(), borrowingRepository, userRepository)
+        val viewModelFactory =
+            ViewModelFactory(ItemRepository(), borrowingRepository, userRepository)
         viewModel = ViewModelProvider(this, viewModelFactory)[ItemViewModel::class.java]
 
         viewModel.items.observe(viewLifecycleOwner) { itemList ->
             adapter.updateItems(itemList)
+        }
+
+        viewModel.summaryItems.observe(viewLifecycleOwner) { itemList ->
+            val summaryMap = itemList
+                .groupingBy { it.item_type }
+                .eachCount()
+                .map { ItemSummary(it.key, it.value) }
+
+            summaryAdapter = ItemSummaryAdapter(summaryMap)
+            binding.rvItemSummary.layoutManager = GridLayoutManager(requireContext(), 2)
+            binding.rvItemSummary.adapter = summaryAdapter
         }
 
         viewModel.loadItems()
@@ -96,14 +121,38 @@ class ItemFragment : Fragment() {
     }
 
     private fun deleteItem(item: Item) {
-        viewModel.deleteItem(item)
-        if (!isDialogVisible) {
-            isDialogVisible = true
-            CustomDialog.alert(
-                context = requireContext(),
-                message = "Item deleted successfully",
-                onDismiss = { isDialogVisible = false }
-            )
+        if (isDialogVisible) return
+
+        isDialogVisible = true
+        CustomDialog.confirm(
+            context = requireContext(),
+            message = "Yakin ingin menghapus item ini?",
+            onConfirm = {
+                viewModel.deleteItem(item)
+                CustomDialog.alert(
+                    context = requireContext(),
+                    message = "Item berhasil dihapus",
+                    onDismiss = { isDialogVisible = false }
+                )
+            },
+            onCancel = {
+                isDialogVisible = false
+            }
+        )
+    }
+
+    private fun setupSwipeToRefresh() {
+        binding.swipeRefresh.setOnRefreshListener {
+            binding.etSearch.setText("")
+            binding.etSearch.clearFocus()
+
+            val imm =
+                requireContext().getSystemService(android.content.Context.INPUT_METHOD_SERVICE) as android.view.inputmethod.InputMethodManager
+            imm.hideSoftInputFromWindow(binding.etSearch.windowToken, 0)
+
+            viewModel.loadItems()
+
+            binding.swipeRefresh.isRefreshing = false
         }
     }
